kis_backend/requirements.txt
Django
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
websocket-client
dotenv

